refactor: extract shared quota-event emit helpers
requestQuotas and readQuotas were duplicated across usage-status,
quota-warnings, and sub-bar-integration. Move the pi.events.emit
glue into extensions/_shared/quota-events.ts so src/ stays free of
pi-coupled code. Payload types remain in src/types/quotas.ts.
